requirements.txt
pandas
numpy
requests
certifi
yfinance
beautifulsoup4
trafilatura
sqlalchemy
psycopg2-binary
python-dotenv
agno>=1.0.0
groq>=0.11.0
pydantic>=2.6.0
ddgs>=9.0.0
